Fix base url in header
When the blog is placed in a sub-folder,  base url is already presented. So don't need to append the base url again, otherwise it will be duplicated.

For example, when place the blog inside another app's folder `blog`, we will set the `base.url` to `/blog/` in `.vuepress/config.js`. Thus, the header url will become `/blog/blog` which is incorrect.